requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-filter==23.2
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
duplicity==0.8.21
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.9
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.45.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.5.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyTelegramBotAPI==4.9.0
python-apt==2.4.0+ubuntu1
python-debian==0.1.43+ubuntu1.1
pytube==12.1.3
pytz==2023.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
tk==0.1.0
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unity-tweak-tool==0.0.7
uritemplate==4.1.1
urllib3==2.0.2
usb-creator==0.3.7
wadllib==1.3.6
websockets==9.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.8.2
yt-dlp==2022.4.8
zipp==1.0.0
